Workers/tests/submissions/Microsoft/WorkerGlobalScope_ErrorEvent_type.htm

<!DOCTYPE html>
<html>
<head>
<title> WorkerGlobalScope ErrorEvent.type </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    setup({timeout:500});
    var ErrorMessage = "Error Message";
    var FileName = './support/ErrorEvent.js';
  
    var t = async_test("Test Description: ErrorEvent and ErrorEvent.type.");
                  
    var worker = new Worker(FileName);

    worker.onmessage = t.step_func(function(evt)
    {
        var err = evt.data;
        assert_array_equals([err.event, err.type], ["[object ErrorEvent]", "error"]);
        t.done();        
    });
    
    worker.postMessage(ErrorMessage);
</script>
</body>
</html>
